requirements.txt
flask
json
mysql.connector-python
